Add files via upload
 This is licensed media of a copyrighted computer game.
Type: Skill-specific icon
This file is taken from RuneScape.com, and its copyright is held by Jagex Ltd. and is used with permission.
This is licensed media of a copyrighted computer game.
Type: Sprite
This file is taken from RuneScape.com, and its copyright is held by Jagex Ltd. and is used with permission.